🍊 Extract Icon enum to separate file in constants package
