+++ app/src/main/java/com/example/sunrise/Register.java
@@ -16,13 +16,13 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
+import com.example.sunrise.constants.Icon;
 import com.example.sunrise.models.Category;
 import com.example.sunrise.models.Tag;
 import com.example.sunrise.models.UserSettings;
 import com.example.sunrise.services.CategoryService;
 import com.example.sunrise.services.TagService;
 import com.example.sunrise.services.UserSettingsService;
-import com.example.sunrise.utils.IconPickerDialog;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.android.material.textfield.TextInputEditText;
@@ -134,7 +134,7 @@ public class Register extends AppCompatActivity {
         UserSettingsService userSettingsService = new UserSettingsService();
 
         // Create a default category
-        Category defaultCategory = new Category("Personal", -13057, IconPickerDialog.Icon.FLOWER.toString(), defaultTagId, user.getUid());
+        Category defaultCategory = new Category("Personal", -13057, Icon.FLOWER.toString(), defaultTagId, user.getUid());
 
         // Save the default category to Firebase
         categoryService.saveCategory(defaultCategory, new OnCompleteListener<Void>() {
